[HAPI-184] Create the CRUD class instance
## [HAPI-184] Description
Created a class to manage the CRUD operations
