+++ src/controllers/collection/get/query.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 const models = require('4hands-api/src/models');
-const CRUD = require('4hands-api/src/services/database/crud');
+const CRUD = require('4hands-api/src/services/Database/CRUD');
 const GetQueryCollection = models.routes.collection.GetQueryCollection;
 const Endpoint = require('4hands-api/src/models/settings/Endpoint');
 const Response = GetQueryCollection.response;
