+++ src/models/DataDependency.js
@@ -1,4 +1,4 @@
-const CRUD = require('4hands-api/src/services/database/crud');
+const CRUD = require('4hands-api/src/services/Database/CRUD');
 
 /**
  * Represents a data dependency used to manage real-time data updates and rendering in a parent component.
