+++ src/services/ServerIO/SubscriptionIO.js
@@ -1,5 +1,5 @@
 const crypto = require('crypto');
-const CRUD = require('4hands-api/src/services/database/crud');
+const CRUD = require('4hands-api/src/services/Database/CRUD');
 
 /**
  * Class representing a SubscriptionIO for managing query and document subscriptions.
